Change default transform type to affine in Registrator class
The default transform type, similarity, was giving strange
behaviour when more than two patches were selected to register the
images. In some cases, it appears that one of the points would be
ignored.

The reason for this behaviour is not clear but could be due to the
estimation method used in scikit-image.transform.estimate_transform
and how it regularises unstable points.