Select the transformation type based on the number of windows
The type of transform was fixed as affine, leading to nonsense results
if fewer than three focus windows were selected. The transform is now
chosen to exactly match the DOF specified by the number of windows if
four or fewer points are chosen, or using projective transformation 
and the least square fit from skimage's estimate transform for more
than four points.